Add server statefulSet.
The server statefulSet is a prerequisite for the boostrapTrustBundle
being set to "path", as it needs to write the configMap that will be
sourced by the agent.

Add server configMap.

The server configMap is a prerequisite for the server StatefulSet as the
server needs a configmap to run.

Server image support added, as it exists in agent image support.
Server configmap support added, as it exists in agent configmap support.
Improved image testing for image, server, and agent with server and
agent overrides.
Makefile modified to include dumping of tested keys for easier
CONFIGURATION.md updates and values.schema.yaml updates.
values.schema.yaml updated

Signed-off-by: Edwin Buck <edwbuck@gmail.com>